# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BrinnPay API
  version: 1.0.0
  description: |
    BrinnPay is a developer-focused payment infrastructure sandbox. It simulates
    payment infrastructure (payments, refunds, idempotency, webhooks, logs)
    without processing real money.

    This document is the canonical, single source of truth for the BrinnPay `v1`
    API surface — `docs/openapi.yaml` (phase 1 specification §13.2). It covers
    the full MVP resource surface: authentication, organizations, projects,
    API keys, customers, payments, refunds, webhooks, idempotency, and logs.

    Conventions fixed here bind every endpoint-level phase:

    - Base path and versioning: `/api/v1` (ADR-0005). No breaking changes within
      a version; the version is pinned per request by the URL. `info.version` is
      the OpenAPI document version (`1.0.0` = `v1` + document revision), not the
      product release (`v0.1.0`, roadmap Phase 26) — see `api-conventions.md`.
    - Naming: plural nouns (`/customers`, `/payments`); nested resources where
      ownership is strict (`/payments/{payment_id}/refunds`); `snake_case` path
      parameters, query parameters, and JSON fields.
    - Authentication (ADR-0006, phase 1 §7.3): API access via
      `Authorization: Bearer <api_key>` where the key is `sk_test_...` /
      `sk_live_...`. Dashboard/user endpoints authenticate via session (JWT
      access token). The two modes are **not equivalent** (API keys scope to one
...
      operation.
...

...
servers:
...
security: []
paths:
...
  /projects/{project_id}:
    get:
      tags: [projects]
      operationId: projects.retrieve
      summary: Retrieve a project
      description: |
        Returns a project. Access requires membership in the owning organization
        and project access within it (phase 1 §11.4); non-members receive 404 so
        project existence is not disclosed.
      security:
        - SessionAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project details.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
    patch:
      tags: [projects]
      operationId: projects.update
      summary: Update a project
      description: |
        Updates project attributes. Requires an administrative role in the
        owning organization.
      security:
        - SessionAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        $ref: "#/components/requestBodies/ProjectUpdateBody"
      responses:
        "200":
          description: Updated project.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/BusinessRuleViolation"
        "429":
          $ref: "#/components/responses/RateLimited"
    delete:
      tags: [projects]
      operationId: projects.delete
      summary: Delete a project
      description: |
        Deletes a project and its tenant-scoped data, including the project's API
        keys (they are removed with the project; the cascade also applies when the
        owning organization is deleted, D7). Requires an administrative
        role in the owning organization.
      security:
        - SessionAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "204":
          description: Project deleted.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/BusinessRuleViolation"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /projects/{project_id}/webhook-endpoints:
    get:
      tags: [webhooks]
      operationId: webhooks.listEndpoints
      summary: List webhook endpoints of a project
      description: |
        Lists the registered webhook endpoints of a project environment with
        cursor pagination. `signing_secret` is never included. Creating or
        enabling an endpoint also delivers the events it missed while it was
        disabled, because the reconciliation pass recomputes the fan-out of
        retained events within a bounded window: the pass scans events from
        `now - WEBHOOK_RECONCILIATION_HORIZON_MS` (1 hour by default) up to
        30 seconds old, and the horizon is rejected at boot if it is not longer
        than that minimum age. Events older than the horizon stay listable and
        replayable through the replay operation, but the safety net does not
        backfill them on its own.
      security:
        - ApiKeyAuth: []
        - SessionAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/Environment"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Paginated webhook endpoint list.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpointList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
    post:
      tags: [webhooks]
      operationId: webhooks.createEndpoint
      summary: Register a webhook endpoint
      description: |
        Registers a destination URL with an event subscription for a project
        environment. The signing secret is returned only once, at creation; it is
        used to verify HMAC-SHA256 signatures of outbound deliveries.

        `event_types` is validated against the closed catalog
        (`payment.created`, `payment.succeeded`, `payment.failed`,
        `refund.created`); it must be a non-empty subset of it, and an unknown
        type is a 400 `VALIDATION_ERROR`, as is a missing or empty `event_types`.
        `url` must be an absolute `http`/`https` URL that satisfies the
        destination policy; malformed and non-HTTP URLs are a 400
        `VALIDATION_ERROR`. With API-key authentication the body's `environment`
        must match the key's environment; an explicit conflict is a 422
        `BUSINESS_RULE_VIOLATION` (an endpoint of the key's project in the other
        environment, addressed by id, is instead a 404 that discloses nothing).
        Duplicate URLs are allowed: there is no uniqueness constraint on
        `(project, environment, url)`.

        The returned `signing_secret` is unrecoverable afterwards (no rotation
        operation in this version); recovery is delete and recreate.
      security:
        - ApiKeyAuth: []
        - SessionAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        $ref: "#/components/requestBodies/WebhookEndpointCreateBody"
      responses:
        "201":
          description: Webhook endpoint created. The signing secret is returned only here.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpointCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/BusinessRuleViolation"
        "429":
          $ref: "#/components/responses/RateLimited"
  /projects/{project_id}/webhook-endpoints/{endpoint_id}:
    get:
      tags: [webhooks]
      operationId: webhooks.retrieveEndpoint
      summary: Retrieve a webhook endpoint
      description: Returns a registered webhook endpoint. The signing secret is never returned.
      security:
        - ApiKeyAuth: []
        - SessionAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/EndpointId"
      responses:
        "200":
          description: Webhook endpoint details.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpoint"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
    patch:
      tags: [webhooks]
      operationId: webhooks.updateEndpoint
      summary: Update a webhook endpoint
      description: |
        Updates the URL, event subscription, or enabled state of a webhook
        endpoint. Omitted fields are unchanged; an empty body is a 400. The
        signing secret is never rotated or returned.

        Setting `enabled` to `false` stops **new** deliveries from being created;
        deliveries already queued for the endpoint still run (D11). A disabled
        endpoint also no longer accepts a replay (422).

        The same validation rules as `webhooks.createEndpoint` apply to `url` and
        `event_types`, and they are reported the same way: a 400
        `VALIDATION_ERROR`. The update body carries no `environment` — it is
        derived from the addressed endpoint and cannot change — so the declared
        422 is reserved and is not produced in this version.
      security:
        - ApiKeyAuth: []
        - SessionAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/EndpointId"
      requestBody:
        $ref: "#/components/requestBodies/WebhookEndpointUpdateBody"
      responses:
        "200":
          description: Updated webhook endpoint.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpoint"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/BusinessRuleViolation"
        "429":
          $ref: "#/components/responses/RateLimited"
    delete:
      tags: [webhooks]
      operationId: webhooks.deleteEndpoint
      summary: Delete a webhook endpoint
      description: |
        Deletes a webhook endpoint, stopping deliveries to it. Its delivery rows
        are removed with it; a delivery job still queued for a deleted endpoint
        is a silent no-op. Requires an administrative role (owner or admin) with
        a session, or an API key scoped to the endpoint's project and
        environment.

        The `422` response below is reserved for contract symmetry and is not
        produced in this version: deletion has no business-rule precondition.
      security:
        - ApiKeyAuth: []
        - SessionAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/EndpointId"
      responses:
        "204":
          description: Webhook endpoint deleted.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/BusinessRuleViolation"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /projects/{project_id}/webhook-endpoints/{endpoint_id}/events/{event_id}/replay:
    post:
      tags: [webhooks]
      operationId: webhooks.replayEvent
      summary: Replay a webhook event to an endpoint
      description: |
        Re-creates a delivery of a retained event to a webhook endpoint as a new
        delivery with the original envelope; the event is not modified and the
        attempt is repeated for the same `(event, endpoint)` pair, so replaying
        twice produces two deliveries (`is_replay` is `true`).

        The event must belong to the same project **and** environment as the
        endpoint; a foreign, cross-environment, or retention-expired event is a
        404 and reveals nothing about either resource. An endpoint that is
        disabled or not subscribed to the event's type is a 422.

        Replay is intentionally repeatable: the partial unique index that
        guarantees at most one fan-out delivery per `(event, endpoint)` excludes
        replays. Requires an administrative role (owner or admin) with a session,
        or an API key scoped to the endpoint's project and environment.
      security:
        - ApiKeyAuth: []
        - SessionAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/EndpointId"
        - $ref: "#/components/parameters/EventId"
      responses:
        "202":
          description: Replay accepted and queued for delivery.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/BusinessRuleViolation"
        "429":
          $ref: "#/components/responses/RateLimited"
  /projects/{project_id}/webhook-events:
    get:
      tags: [webhooks]
      operationId: webhooks.listEvents
      summary: List webhook events of a project
      description: |
        Lists the outbound events emitted in a project environment with cursor
        pagination. Each row is the stored outbound envelope
        (`id`, `type`, `environment`, `project_id`, `data`, `created_at`). Events
        are retained for a configurable window (default 30 days); an expired
        event is removed by the cleanup pass, absent here, and 404 on replay.
      security:
        - ApiKeyAuth: []
        - SessionAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/Environment"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/EventType"
      responses:
        "200":
          description: Paginated webhook event list.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEventList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: sk_test_... | sk_live_...
      description: |
        API-key authentication for programmatic access (ADR-0006). The key is
        sent as `Authorization: Bearer sk_test_...` (TEST) or
        `Authorization: Bearer sk_live_...` (LIVE). A key is scoped to exactly
        one project environment and can only act within that scope; it never
        carries organization-management authority. Keys are shown once at
        creation, stored hashed, and never logged.
    SessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Session authentication with a short-lived JWT access token for
        dashboard/user endpoints (phase 1 §7.3, §11.5). Tokens are obtained
        through `/auth/login` and refreshed through `/auth/refresh` (refresh
        token travels only as an `HttpOnly` cookie). Session requests are scoped
        to the authenticated user's organization memberships and subject to RBAC.
  headers:
    X-Request-Id:
      description: |
        Request ID assigned to the request at ingress (phase 1 §7.7). Every
        response carries it; it is propagated through logs, error responses, and
        outbound webhook delivery records.
      schema:
        type: string
        example: req_0192f2a0-0000-7000-8000-000000000001
...
  parameters:
...
    ProjectId:
      name: project_id
      in: path
      required: true
      description: Project UUIDv7 ID, treated as an opaque string (ADR-0001).
      schema:
        type: string
        format: uuid
        example: 0192f2a0-0000-7000-8000-00000000000b
...
    EndpointId:
      name: endpoint_id
      in: path
      required: true
      description: Webhook endpoint UUIDv7 ID, treated as an opaque string (ADR-0001).
      schema:
        type: string
        format: uuid
        example: 0192f2a0-0000-7000-8000-000000000005
    EventId:
      name: event_id
      in: path
      required: true
      description: Webhook event UUIDv7 ID, treated as an opaque string (ADR-0001).
      schema:
        type: string
        format: uuid
        example: 0192f2a0-0000-7000-8000-000000000006
...
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of results per page (cursor pagination, phase 1 §7.5).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      required: false
      description: |
        Opaque cursor returned as `next_cursor` in the previous page. The server
        derives cursors from the time-ordered entity IDs (UUIDv7, ADR-0001).
        Clients must not decode cursors or reuse them across unrelated lists.
      schema:
        type: string
...
    Environment:
      name: environment
      in: query
      required: false
      description: |
        Project environment filter (`test` or `live`). With session
        authentication this selects the environment to operate on. With API-key
        authentication the environment is derived from the key and must match.
      schema:
        $ref: "#/components/schemas/Environment"
    EventType:
      name: type
      in: query
      required: false
      description: |
        Optional event-type filter for `webhooks.listEvents` (D15). Must be a
        member of the closed webhook catalog; an unknown type is a 400.
      schema:
        type: string
        enum:
          [
            payment.created,
            payment.succeeded,
            payment.failed,
            refund.created,
          ]
        example: payment.succeeded
...
  requestBodies:
...
    ProjectUpdateBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProjectUpdate"
...
    WebhookEndpointCreateBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/WebhookEndpointCreate"
    WebhookEndpointUpdateBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/WebhookEndpointUpdate"
  responses:
    BadRequest:
      description: Validation failed — the request is malformed or a field violates validation rules.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/X-Request-Id"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthenticated:
      description: Authentication is missing or invalid.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/X-Request-Id"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: The authenticated principal is not authorized for this operation (RBAC or tenant isolation).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/X-Request-Id"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: The requested resource does not exist or is not visible to the caller.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/X-Request-Id"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    BusinessRuleViolation:
      description: |
        A domain/business rule was violated (e.g., refunding an already fully
        refunded payment, `PAYMENT_ALREADY_REFUNDED`).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/X-Request-Id"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimited:
      description: The request exceeded an applicable rate limit.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/X-Request-Id"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    Environment:
      type: string
      enum: [test, live]
      description: |
        Project API environment (phase 1 §5.2). Both environments are simulated;
        no real money is processed in either. TEST and LIVE data is never mixed.
...
    ErrorEnvelope:
      type: object
      required: [error]
      description: |
        Consistent JSON error envelope (phase 1 §7.6). `code` is a stable
        machine-readable string (general codes such as `VALIDATION_ERROR`,
        `UNAUTHENTICATED`, `FORBIDDEN`, `NOT_FOUND`, `CONFLICT`, `RATE_LIMITED`,
        plus business-specific codes such as `PAYMENT_ALREADY_REFUNDED`).
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              description: Stable, machine-readable error code.
              example: PAYMENT_ALREADY_REFUNDED
            message:
              type: string
              description: Human-readable error message. Never leaks internals, stack traces, or secrets.
              example: This payment has already been fully refunded.
            request_id:
              type: string
              description: Request ID assigned at ingress (phase 1 §7.7).
              example: req_0192f2a0-0000-7000-8000-000000000001
            details:
              type: object
              additionalProperties: true
              description: Optional structured information (e.g., validation field errors).
...
    Project:
      type: object
      required: [id, organization_id, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        environments:
          type: array
          items:
            $ref: "#/components/schemas/Environment"
          description: |
            The API environments available for the project. A project supports
            both `test` and `live` by default; each environment-scoped resource
            belongs to exactly one environment and TEST/LIVE data is never mixed
            (phase 1 §5.2).
...
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
...
    WebhookEndpoint:
      type: object
      required:
        [
          id,
          project_id,
          environment,
          url,
          event_types,
          enabled,
          created_at,
          updated_at,
        ]
      description: |
        A developer-registered destination URL with an event subscription. The
        signing secret is never returned by the API.
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        environment:
          $ref: "#/components/schemas/Environment"
        url:
          type: string
          format: uri
        event_types:
          type: array
          items:
            $ref: "#/components/schemas/WebhookEventType"
          description: Subscribed event types from the closed webhook catalog.
        enabled:
          type: boolean
          description: |
            When `false`, no new delivery is created for this endpoint. A
            delivery already queued still runs (D11).
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WebhookEndpointCreated:
      type: object
      required:
        [
          id,
          project_id,
          environment,
          url,
          event_types,
          enabled,
          signing_secret,
          created_at,
          updated_at,
        ]
      description: |
        Webhook endpoint creation response. The `signing_secret` is returned
        exactly once and is used to verify HMAC-SHA256 signatures of outbound
        deliveries. The secret is never logged or returned again.
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        environment:
          $ref: "#/components/schemas/Environment"
        url:
          type: string
          format: uri
        event_types:
          type: array
          items:
            $ref: "#/components/schemas/WebhookEventType"
        enabled:
          type: boolean
        signing_secret:
          type: string
          description: |
            Endpoint signing secret, returned only at creation. It is stored
            encrypted (AES-256-GCM) and is **not** recoverable through the API:
            there is no rotation or reveal operation, so it must be copied at
            creation time. Deleting and recreating the endpoint is the only way
            to obtain a new secret (D8).
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    WebhookEndpointCreate:
      type: object
      required: [environment, url, event_types]
      additionalProperties: false
      properties:
        environment:
          $ref: "#/components/schemas/Environment"
          description: |
            The project environment the endpoint receives events from. With
            API-key authentication it must match the key's environment.
        url:
          type: string
          format: uri
          maxLength: 2048
          description: |
            Absolute destination URL. Only `http` and `https` are accepted, and
            the URL must satisfy the server's destination policy. Redirects are
            never followed, so the destination must answer directly.
        event_types:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/WebhookEventType"
        enabled:
          type: boolean
          default: true
...
    WebhookEndpointList:
      type: object
      required: [data, next_cursor, has_more]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/WebhookEndpoint"
        next_cursor:
          type: [string, "null"]
        has_more:
          type: boolean
    WebhookEventList:
      type: object
      required: [data, next_cursor, has_more]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/WebhookEvent"
        next_cursor:
          type: [string, "null"]
        has_more:
          type: boolean
...